These components seem to be working
As a first step in making my FPGA interface with a monitor over VGA, I've made some basic counters, a clock divider, and a logic unit for detecting when to send sync pulses. The stripes_tb file is a testbench which should make stripes of various colors on the screen when run on an FPGA, and it seems to work in Vivado, but the output waveform isn't a very fun way of verifying that the code I've made so far works.