Horizontal stripes and a ROM added
Vertical and horizontal test patterns are working. The ROM is working as well, but will need to be changed so that each pixel on the screen has its own address in the memory (or, two pixels could be stored in a single byte of the memory) and each pixel can be addressed. Currently each pixel cannot be addressed as the address is only 12 bits long whereas the horizontal+vertical beam counter sizes are 17 bits long total (currently).
The picture appears to be a little blurry when the value of pixels changes based on the h_counter (like when displaying horizontal stripes), which might just be a limitation of the 10MHz clock that I've set the driver to run at.
Also, I forgot to/still have not divided the number of vertical lines by 4 to compensate for the reduction in clock speed. This doesn't seem to stop the driver from working, but it might be why the monitor says that the display is in 800x600@30Hz.