Added basic framebuffer support
Each new line of the screen should be 32 bits apart (in hex: 00, 20, 40, 60, etc). There are 72 lines on the screen, with each line having 25 pixels (so from +0 to +18 offset-wise). I resized my Quartus II mif editor window such that the row offset part went from +0 to +1f and then changed the window size such that only +0 to +18 were in view. This gave me a pretty good space to manually edit the 3-bit-words which each correspond to a horizontal pixel-like-think on the screen.
I also decided to use one of the PLLs to get 40MHz so I could do proper 800x600.