Hello, World! :)
Hello should be in white, world should be in blue (the 'W' is a bit messed up), and the exclamation point should be red. I put some cyan and all the colors at the bottom of the screen to demonstrate the boundaries of the display. 
Note: the framebuffer seems to have a very slight warm-up period at the beginning of the display where it will display the old value (from the end of the last line).