Memory address test
Used to test the range of memory addresses accessed by the counters. To use this test bench, you must make some changes to the vga driver so that it outputs the address that is fed into the ROM. I only used this particular test bench to find out the range of addresses accessed--I usually just test the vga code on the FPGA, since it's easier to see pixels on a display as opposed to within a waveform.