Square pixels, memory address usage reduced
I squared off the pixels and, using some multiplying and bit-shifting, altered the address space of the frame buffer so that it is all in one continuous block of memory as opposed to scattered chunks of memory (portions of the memory would end up unused due to the blanking period not using pixels when the counters were directly mapped to memory addresses).
I had quite a bit of problems regarding the image being vertically shifted down a few lines, which cut off some lines at the bottom of the screen. I was able to use an oscilloscope to verify that this was just a problem with my monitor's settings, and the problem was resolved by resetting my monitor.
One minor issue that still remains is the small bit of out-of-bounds memory that is accessed on the 0th line of the screen--this is due to the timing of the setting of the memory mapping function (set on the falling edge of the HSYNC) which technically places the address one pixel in the past.